+++ src/a2p.Shared/Application/DTO/MaterialDTO.cs
@@ -25,34 +25,34 @@ namespace a2p.Shared.Application.DTO
         [MaxLength(50)] public string Color { get; set; } = string.Empty;
         [MaxLength(120)] public string? ColorDescription { get; set; }
         //============================================================================================================================
-        public double Width { get; set; } = 0;
-        public double Height { get; set; } = 0;
+        public decimal Width { get; set; } = 0m;
+        public decimal Height { get; set; } = 0m;
         //============================================================================================================================       
         [Required] public int Quantity { get; set; } = 0;
-        public double PackageQuantity { get; set; } = 0;
-        public double TotalQuantity { get; set; } = 0;
-        [Required] public double RequiredQuantity { get; set; } = 0;
-        public double LeftOverQuantity { get; set; } = 0;
+        public decimal PackageQuantity { get; set; } = 0m;
+        public decimal TotalQuantity { get; set; } = 0m;
+        [Required] public decimal RequiredQuantity { get; set; } = 0m;
+        public decimal LeftOverQuantity { get; set; } = 0m;
         //============================================================================================================================  
-        public double Weight { get; set; } = 0;
-        public double TotalWeight { get; set; } = 0;
-        public double RequiredWeight { get; set; } = 0;
-        public double LeftOverWeight { get; set; } = 0;
+        public decimal Weight { get; set; } = 0m;
+        public decimal TotalWeight { get; set; } = 0m;
+        public decimal RequiredWeight { get; set; } = 0m;
+        public decimal LeftOverWeight { get; set; } = 0m;
         //============================================================================================================================  
-        public double Area { get; set; } = 0;
-        public double TotalArea { get; set; } = 0;
-        public double RequiredArea { get; set; } = 0;
-        public double LeftOverArea { get; set; } = 0;
+        public decimal Area { get; set; } = 0m;
+        public decimal TotalArea { get; set; } = 0m;
+        public decimal RequiredArea { get; set; } = 0m;
+        public decimal LeftOverArea { get; set; } = 0m;
         //============================================================================================================================  
-        public double Waste { get; set; } = 0;
+        public decimal Waste { get; set; } = 0m;
         //============================================================================================================================  
-        public double Price { get; set; } = 0;
-        public double TotalPrice { get; set; } = 0;
-        public double RequiredPrice { get; set; } = 0;
-        public double LeftOverPrice { get; set; } = 0;
+        public decimal Price { get; set; } = 0m;
+        public decimal TotalPrice { get; set; } = 0m;
+        public decimal RequiredPrice { get; set; } = 0m;
+        public decimal LeftOverPrice { get; set; } = 0m;
         //============================================================================================================================  
-        public double SquareMeterPrice { get; set; } = 0;
-        //============================================================================================================================
+        public decimal SquareMeterPrice { get; set; } = 0m;
+        //============================================================================================================================`
         [MaxLength(255)] public string? Pallet { get; set; }
         //============================================================================================================================
         [MaxLength(255)] public string? CustomField1 { get; set; }
